pyproject.toml
[tool.poetry]
name = "sf-mmorpg"
version = "0.1.0"
description = ""
authors = ["LatikDesu <evga.abrosimov@gmail.com>"]
readme = "README.md"
packages = [{include = "sf_mmorpg"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-allauth = "^0.54.0"
django-tinymce = "^3.6.1"
django-debug-toolbar = "^4.1.0"
pillow = "^10.0.0"
django-filter = "^23.2"
celery = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
